Create a unified mock thread pool (WIP)
