+++ tests/benchmark/bm_vecsim_basics.h
@@ -62,7 +62,7 @@ void BM_VecSimBasics<index_type_t>::AddLabel(benchmark::State &st) {
     }
     memory_delta = index->getAllocationSize() - memory_delta;
     // For tiered index, wait for all threads to finish indexing
-    BM_VecSimGeneral::thread_pool_wait();
+    tiered_index_mock::thread_pool_wait(BM_VecSimGeneral::jobQ);
 
     st.counters["memory_per_vector"] = (double)memory_delta / (double)added_vec_count;
     st.counters["vectors_per_label"] = vec_per_label;
@@ -78,7 +78,6 @@ void BM_VecSimBasics<index_type_t>::AddLabel(benchmark::State &st) {
         VecSimIndex_DeleteVector(
             INDICES[st.range(0) == VecSimAlgo_TIERED ? VecSimAlgo_HNSWLIB : st.range(0)], label);
     }
-    BM_VecSimGeneral::thread_pool_wait();
     assert(VecSimIndex_IndexSize(index) == N_VECTORS);
 }
 
@@ -106,7 +105,7 @@ void BM_VecSimBasics<index_type_t>::AddLabel_AsyncIngest(benchmark::State &st) {
         added_vec_count += vec_per_label;
         label++;
         if (label == initial_label_count + BM_VecSimGeneral::block_size) {
-            BM_VecSimGeneral::thread_pool_wait();
+            tiered_index_mock::thread_pool_wait(BM_VecSimGeneral::jobQ);
         }
     }
 
@@ -155,7 +154,7 @@ void BM_VecSimBasics<index_type_t>::DeleteLabel(algo_t *index, benchmark::State 
     }
     memory_delta = index->getAllocationSize() - memory_before;
 
-    BM_VecSimGeneral::thread_pool_wait();
+    tiered_index_mock::thread_pool_wait(BM_VecSimGeneral::jobQ);
     // Remove the rest of the vectors that hadn't been swapped yet for tiered index.
     if (VecSimIndex_BasicInfo(index).algo == VecSimAlgo_TIERED) {
         reinterpret_cast<TieredHNSWIndex<data_t, data_t> *>(index)->executeReadySwapJobs();
@@ -171,7 +170,7 @@ void BM_VecSimBasics<index_type_t>::DeleteLabel(algo_t *index, benchmark::State 
             VecSimIndex_AddVector(index, removed_labels_data[label_idx][vec_idx].data(), label_idx);
         }
     }
-    BM_VecSimGeneral::thread_pool_wait();
+    tiered_index_mock::thread_pool_wait(BM_VecSimGeneral::jobQ);
     size_t cur_index_size = VecSimIndex_IndexSize(index);
     assert(VecSimIndex_IndexSize(index) == N_VECTORS);
 }
@@ -201,7 +200,7 @@ void BM_VecSimBasics<index_type_t>::DeleteLabel_AsyncRepair(benchmark::State &st
         // Delete label
         VecSimIndex_DeleteVector(tiered_index, label_to_remove++);
         if (label_to_remove == BM_VecSimGeneral::block_size) {
-            BM_VecSimGeneral::thread_pool_wait();
+            tiered_index_mock::thread_pool_wait(BM_VecSimGeneral::jobQ);
         }
     }
 
@@ -230,7 +229,7 @@ void BM_VecSimBasics<index_type_t>::DeleteLabel_AsyncRepair(benchmark::State &st
                                   label_idx);
         }
     }
-    BM_VecSimGeneral::thread_pool_wait();
+    tiered_index_mock::thread_pool_wait(BM_VecSimGeneral::jobQ);
     assert(VecSimIndex_IndexSize(tiered_index) == N_VECTORS);
 }
